# docs/openapi.yaml
openapi: 3.0.3
info:
  title: نظام إدارة محل السبائك والمشغولات — API
  description: |
    REST API لنظام إدارة محل الذهب والفضة: مخزون مركزي، تسعير يومي، فروع متعددة،
    نقاط بيع (POS)، حجوزات، جرد، شيفتات، تقارير، ومزامنة مع WooCommerce.

    ## المصادقة
    - جميع النقاط تحتاج `Authorization: Bearer <token>` ما عدا `/api/auth/login` و `/api/health`.
    - توكن JWT (خوارزمية HS256، صلاحية 12 ساعة) يُحصل عليه من `/api/auth/login`.
    - لكل مسؤول صلاحيات (permissions) تُحمّل مع كائن الموظف عند الدخول.
      النقاط المحمية بصلاحية محددة موثّقة بـ «Requires permission» في الوصف،
      وترجع `403 {error:"forbidden"}` عند غيابها.

    ## الأخطاء
    جميع الأخطاء تُرجع JSON على الصورة: `{"error": "<code>"}` مع اختياري
    `message` (نص عربي قابل للعرض) و `detail` (لأخطاء الاتصال) و `refs` (للحذف المقيّد).
    الكود الداخلي في `error` ثابت وقابل للبرمجة عليه.

    ## توكن الدخول
    استخدم `/api/auth/login` بالحسابات الأولية (PIN: `1234`):
    | username | الدور |
    |---|---|
    | `manager` | مدير المحل |
    | `cashier` | كاشير |
    | `social` | سوشيال |
  version: 1.0.0
  contact:
    name: Jewelry Store System
servers:
  - url: http://localhost:4001
    description: تطوير (Vite proxy)
  - url: https://your-domain.example
    description: إنتاج
...
security:
  - bearerAuth: []
paths:
...
  /api/locations:
    get:
      tags: [Locations]
      summary: قائمة الفروع (بما فيها غير النشط)
      responses:
        '200':
          description: الفروع
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Location' }
    post:
      tags: [Locations]
      summary: إضافة فرع
      description: "Requires permission: `locations.manage`. يُعيد 409 `duplicate` عند تكرار الكود."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, nameAr]
              properties:
                code: { type: string, description: يُحوَّل لأحرف كبيرة، مثال MAIN }
                nameAr: { type: string }
                nameEn: { type: string }
      responses:
        '201':
          description: الفرع الجديد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/locations/{id}:
    put:
      tags: [Locations]
      summary: تعديل فرع
      description: "Requires permission: `locations.manage`"
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                nameAr: { type: string }
                nameEn: { type: string }
                isActive: { type: boolean }
      responses:
        '200':
          description: الفرع بعد التعديل
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Locations]
      summary: حذف فرع (مع نقل المنتجات/الحركات)
      description: |
        Requires permission: `locations.manage`.
        - 409 `locations.in_use` إذا كان على الفرع موظفون/فواتير/شيفتات/جرد.
        - إذا كان للفرع منتجات أو حركات نقل فيجب تمرير `moveToLocationId`
          ويتم نقل المنتجات وحركات النقل (المتجهة والمنطلقة) إلى الفرع المستهدف.
        - يُعيد `movedItems` و `movedMovements`.
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                moveToLocationId:
                  type: integer
                  description: الفرع المستهدف (مطلوب إذا كان للفرع منتجات أو حركات)
      responses:
        '200':
          description: تم الحذف
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  movedItems: { type: integer, example: 5 }
                  movedMovements: { type: integer, example: 2 }
        '400':
          description: moveTo.required / moveTo.same / moveTo.notfound
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/movements:
    get:
      tags: [Movements]
      summary: حركات النقل (آخر 500)
      parameters:
        - { name: itemId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [in_transit, received, cancelled] } }
      responses:
        '200':
          description: الحركات
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Movement' }
    post:
      tags: [Movements]
      summary: إنشاء حركة نقل بين فروع
      description: |
        Requires permission: `movement.create`.
        تزيد `in_transit_qty` وتصبح القطعة `in_transit` حتى تأكيد الاستلام.
        409 `items.not_available` إذا كانت الكمية المتاحة غير كافية.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId, toLocationId]
              properties:
                itemId: { type: integer }
                toLocationId: { type: integer }
                quantity: { type: integer, default: 1 }
                reason: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '400':
          description: missing / bad.quantity / movements.same_location
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/movements/{id}/receive:
    post:
      tags: [Movements]
      summary: تأكيد استلام النقل (إلزامي لإتمامه)
      description: |
        Requires permission: `movement.receive`.
        ينقل القطعة فعليًا إلى الفرع الوجهة ويعيد حساب الحالة.
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: movements.not_in_transit

  # ------------------------------------------------------------------
  # Stock counts
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توكن JWT من /api/auth/login (صلاحية 12 ساعة)
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: معرّف رقمي
  responses:
...
    NotFound:
      description: غير موجود
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: تعارض (حالة غير صالحة أو مرجع محجوب)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: كود خطأ ثابت قابل للبرمجة، مثال notfound / items.in_use }
        message: { type: string, description: نص عربي قابل للعرض (اختياري) }
        detail: { type: string, description: تفاصيل إضافية (اختياري) }
        refs:
          type: object
          description: إحصاءات المراجع للحذف المقيّد (اختياري)

    # ---- Auth ----
...
    Location:
      type: object
      properties:
        id: { type: integer }
        code: { type: string, example: MAIN }
        nameAr: { type: string }
        nameEn: { type: string, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Movement:
      type: object
      properties:
        id: { type: integer }
        itemId: { type: integer }
        itemCode: { type: string }
        itemName: { type: string }
        quantity: { type: integer }
        fromLocationId: { type: integer, nullable: true }
        fromLocation: { type: string, nullable: true }
        toLocationId: { type: integer }
        toLocation: { type: string }
        movedBy: { type: integer, nullable: true }
        movedByName: { type: string, nullable: true }
        movedAt: { type: string, format: date-time }
        status: { type: string, enum: [in_transit, received, cancelled] }
        receivedBy: { type: integer, nullable: true }
        receivedByName: { type: string, nullable: true }
        receivedAt: { type: string, format: date-time, nullable: true }
        reason: { type: string, nullable: true }

    # ---- Stock counts ----
...